+++ .eslintrc.js
@@ -4,7 +4,6 @@ module.exports = {
   },
   extends: [
     'plugin:vue/recommended',
-    // 'plugin:prettier/recommended',
     '@vue/prettier',
     'eslint:recommended'
   ],