+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.0 (June 13, 2019)
+## 2.0.0 (June 20, 2019)
 
 - BREAKING: Changed default class for results list from `autocomplete-results` to `autocomplete-result-list`
 - BREAKING: Changed default ID for results list from `autocomplete-results-{id}` to `autocomplete-result-list-{id}`
@@ -6,9 +6,13 @@
 
 ### autocomplete-vue
 
-- BREAKING: Updated default slot to provide template for entire component
-- BREAKING: Added new slot called "result" to provide template for a single result item
+- BREAKING: Updated default slot to control template for entire component
+- BREAKING: Added new slot called `result` to control template for a single result item
 - BREAKING: Removed `onSubmit` prop to use custom `submit` event instead
+
+### autocomplete-js
+
+- BREAKING: Removed `renderResults` option and replaced with `renderResult`, which can be used to control rendering of a single result item. This function can return either a DOM element or an HTML string.
 
 ## 1.0.2 (May 1, 2019)